Add bill creation functionality and fix backend issues
This commit:
1. Refactors the bill POST endpoint to use a request body
2. Adds a missing /visits/{id} endpoint in the visits service
3. Fixes ID type inconsistency between services
4. Enhances the VisitDto with smart price determination
5. Adds a bill creation form to the frontend
6. Adds documentation for the billing features

+++ spring-petclinic-billing-service/src/main/java/org/springframework/samples/petclinic/billing/config/VisitClient.java
@@ -8,5 +8,5 @@ import org.springframework.web.bind.annotation.PathVariable;
 @FeignClient(name = "visits-service")
 public interface VisitClient {
     @GetMapping("/visits/{id}")
-    VisitDto getVisitById(@PathVariable("id") Long id);
+    VisitDto getVisitById(@PathVariable("id") Integer id);
 }
